README: Add hint about no XML reports for base
xml_test_output monkey patches openerp.modules.module.run_unit_tests to
generate XML reports for unit tests. It won't do so for the unit tests
in base, because base is loaded before xml_test_output and at that time
the monkey patch isn't in place yet. Note this in the README, because
otherwise people might not realize that they won't get the reports for
many of their tests.