README: Enable XML output for base with --load
Previously I had added a note that xml_test_output won't output XML
reports for the tests in base. A coworker made me aware that this is
wrong. With the --load argument you can tell Odoo to import
xml_test_output first and thus enable XML reports for the tests in base.
Update the README accordingly.